Corrected multiplicative factor in D(x)
I made the following changes:
(1) Changed denominator of leading coefficient of gckkg from π to np.sqrt(π). There is an error in Eq. 23 of Kitamura's work. If you check Kitamura's reference for the equation (Ref 48, Eq. 3), it is correct there. I think you noticed the equation isn't correct because you added the factor of 2 to the D(x) function to compensate. But it actually comes out to about 1.77.
(2) Changed D(x) function to use the scipy implementation of Dawson's integral. It gives the same result as your function but is much faster, assuming you remove the extra factor of 2.